Refactoring for issue #10 done all objetcs use the same recursive procedure sub-objexts are indicated in that procedure - loadObjects only project is handled differently Adding reportingFilters, stages, tasks, widgets
