--- lib/dslCode/issue12/later_personas/serviceDeveloper/persona.groovy
@@ -1,11 +1,0 @@
-persona 'SD12', {
- 
-  personaDetail {
-    personaCategoryName = 'Deployment Automation'
-    personaPage = [
-      'Microservices',
-      'Microservice Deployments',
-    ]
-  }
-  homePageName = 'Microservice Deployments'
-}
