--- lib/dslCode/issue12/projects/issue12/catalogs/testCatalog12/catalog.groovy
@@ -1,4 +1,0 @@
-catalog 'testCatalog12', {
-	description = 'val12'
-	tags = ["arg1", "arg2"]
-}
