--- lib/dslCode/issue12/projects/issue12/catalogs/testCatalog12/catalogItems/testCatalogItem12/catalogItem.groovy
@@ -1,6 +1,0 @@
-def catItemName='testCatalogItem12'
-
-catalogItem catItemName, {
-  description = 'val12'
-  tags = ["arg1", "arg2"]
-}
