--- lib/dslCode/issue12/projects/issue12/environments/testEnv12/clusters/testCluster12/cluster.groovy
@@ -1,9 +1,0 @@
-cluster 'testCluster12', {
-  description = 'val12'
-  tags = ["arg1", "arg2"]
-  pluginKey = null
-  pluginProjectName = null
-  providerClusterName = null
-  providerProjectName = null
-  provisionProcedure = null
-}
