+++ lib/dslCode/issue12/projects/issue12/environments/testEnv12/environmentTiers/Tier12/environmentTier.groovy
@@ -1,0 +1,7 @@
+  environmentTier 'Tier12', {
+    batchSize = null
+    batchSizeType = null
+    resourceName = [
+      'res12',
+    ]
+  }
