--- lib/dslCode/issue12/projects/issue12/procedures/testProcedure/steps/echo.pl
@@ -1,2 +1,0 @@
-
-print "Hello World\n";
