--- lib/dslCode/issue12/resources/res12/properties/prop1
@@ -1,1 +1,0 @@
-val12
