+++ lib/dslCode/victor/projects/SAMPLE_VAL/dashboards/testDashboard/reportingFilters/DateFilter/reportingFilter.groovy
@@ -1,0 +1,9 @@
+reportingFilter 'DateFilter', {
+  operator = 'BETWEEN'
+  orderIndex = '1'
+  parameterName = '@timestamp'
+  reportObjectTypeName = null
+  required = '1'
+  type = 'DATE'
+  widgetName = null
+}
