+++ specs/src/test/groovy/com/electriccloud/plugin/spec/issue12.groovy
@@ -12,7 +12,6 @@ class issue12 extends PluginTestHelper {
     plugDir = getP("/server/settings/pluginsDirectory")
     dsl """
       deleteProject(projectName: "$projName")
-      deletePersona(personaName: "serviceDeveloper")
       deleteResource(resourceName: "res12")
     """
   }
@@ -20,7 +19,6 @@ class issue12 extends PluginTestHelper {
   def doCleanupSpec() {
     conditionallyDeleteProject(projName)
     dsl """
-      deletePersona(personaName: "serviceDeveloper")
       deleteResource(resourceName: "res12")
    """
   }
@@ -41,49 +39,6 @@ class issue12 extends PluginTestHelper {
     then: "job succeeds"
       assert p.jobId
       assert getJobProperty("outcome", p.jobId) == "success"
-
-     // Issue #2 - persona
- /*   then: "persona is found"
-      def pa=dsl """getPersona(personaName: 'SD12')"""
-*/
-
-    // check resource is found
-    then: "resource is found"
-      def rsc=dsl """
-        getResource(
-          resourceName: "res12"
-        )"""
-      assert rsc.resource.resourceName == "res12"
-      assert rsc.resource.hostName == 'doesnotexist12'
-      assert getP("/resources/res12/prop1") =~ /val12\s+/
-
-    // check resource is found
-    then: "resource nestedproperty is found"
-      assert getP("/resources/res12/level1/level2/level3/prop123") =~ /res3-level3\s+/
-
-    // check service is found
-    then: "service is found"
-      def serv=dsl """
-        getService(
-          projectName: "$projName",
-          serviceName: "service12"
-        )"""
-      assert serv.service.description == "desc12"
-
-     // Check catalog exist
-     then: "catalog exist"
-       println "Checking catalog"
-       assert getP("/projects/$projName/catalogs/testCatalog12/description") == "val12"
-
-     // catalogItem is created
-     then: "catalogItem exist"
-       println "Checking catalog item"
-       def item=dsl """getCatalogItem(
-           projectName: "$projName",
-           catalogName: 'testCatalog12',
-           catalogItemName: 'testCatalogItem12'
-         )"""
-       assert item.catalogItem.description == 'val12'
 
      // check master component
      then: "master component exist"
@@ -110,19 +65,11 @@ class issue12 extends PluginTestHelper {
         getEnvironmentTier(
           projectName: "$projName",
           environmentName: "testEnv12",
-          environmentTierName: "Tier 12",
+          environmentTierName: "Tier12",
         )"""
-      assert envT.environmentTier.environmentTierName == "Tier 12"
+      assert envT.environmentTier.environmentTierName == "Tier12"
     // Check for cluster
-    then: "cluster is found"
-      def cl = dsl """
-        getCluster(
-          projectName: "$projName",
-          environmentName: "testEnv12",
-          clusterName: 'testCluster12',
-        )"""
-      assert cl.cluster.clusterName == "testCluster12"
-      assert cl.cluster.description == "val12"
+
 
     // check release is found
     then: "release is found"
